ml_model/requirements.txt
numpy==1.26.3
pandas==2.1.4 # Bug with 2.2.0 and SQLAlchemy: https://github.com/pandas-dev/pandas/issues/57049
scikit-learn==1.4.0
sqlalchemy==2.0.28
mlflow==2.11.0
python-dotenv==1.0.1
Flask==2.2.5